Update the Whole Code
The whole Code is modified.
The Code and the Algorithm is made much Simpler, and perhaps more efficient.
Most of the Variables are removed, and some are renamed (according to the Need)
The main Algorithm and Technique remains the Same, just modified a little.
The Code Files are merged, (No Need for the Separate File).
Some unintentional Code is removed (which made no effect on the overall Algorithm)

The Interface is now Cleaner and more User-Friendly.
The "How to Play" is now mentioned.
It now uses X and O for Player 1 and 2 respectively (instead of 1 and 2 which doesn't look good).